Notify in Slack on GitHub webhook
Adds logic for parsing incoming events from GitHub and sends a Slack Message to the targeted users. These messages are : ReviewRequested (when a user is set as reviewer for a PR), ReviewSubmitted (when a review is submitted to a user's PR), TaggedInIssue (when mentioned in an issue/PR description), TaggedInIssueComment (when mentioned in an issue/PR comment).